requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
schedule==1.2.0
python-dotenv==1.0.0
pandas==2.1.4
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
